Bypass ko<->React adapter in SettingsTab
